+++ authority-portal-frontend/src/app/core/api/fake-backend/impl/organization-details-fake.ts
@@ -1,10 +1,10 @@
-import {OrganizationDetailResult} from '@sovity.de/authority-portal-client';
+import {OwnOrganizationDetailsDto} from '@sovity.de/authority-portal-client';
 import {TEST_ORGANIZATIONS} from './fake-organizations';
 
 export const getOrganizationDetails = (
   mdsId: string,
-): OrganizationDetailResult => {
+): OwnOrganizationDetailsDto => {
   return TEST_ORGANIZATIONS.find(
     (organization) => organization.mdsId === mdsId,
-  ) as OrganizationDetailResult;
+  ) as OwnOrganizationDetailsDto;
 };