+++ CHANGELOG.md
@@ -24,8 +24,8 @@ Initial Authority Portal MVP Release with a minimalistic feature set.
 - Unregister own connectors
 - Added userId to /api/user-info endpoint
 - Added endpoint to retrieve user data
-
 - Added page to invite new users to an existing organization.
+- Added endpoint to retrieve organization details
 
 #### Patch